feat: restructure Next.js project layout and remove unused files
