+++ .github/workflows/deploy.yml
@@ -29,8 +29,8 @@ jobs:
             npm install
           fi
 
-      - name: Build and export
-        run: npm run export
+      - name: Build
+        run: npm run build
 
       - name: Deploy to gh-pages
         uses: peaceiris/actions-gh-pages@v3
